Fix RSQ instruction error by replace the long data with integer data in the Fast Inverse Square Root implementation

+++ src/common.h
@@ -410,16 +410,19 @@ inline const floatVec4 fvInt2Float(const floatVec4 &x)
 /**
  *	Fast inverse square root
  *	Reference: http://en.wikipedia.org/wiki/Fast_inverse_square_root
+ *  Attention!! the variable "i" needs to be declared as integer. Some
+ *  compiler has different definition in long type and make the result
+ *  comprimised.
  */
 inline const float Q_rsqrt(float number)
 {
-	long i;
+	int i;
 	float x2, y;
 	const float threehalfs = 1.5F;
 
 	x2 = number * 0.5F;
 	y  = number;
-	i  = * ( long * ) &y;                       // evil floating point bit level hacking
+	i  = * ( int * ) &y;                       // evil floating point bit level hacking
 	i  = 0x5f3759df - ( i >> 1 );               // what the fuck?
 	y  = * ( float * ) &i;
 	y  = y * ( threehalfs - ( x2 * y * y ) );   // 1st iteration
